+++ packages/react-hooks/src/index.ts
@@ -3,8 +3,11 @@ export type {
   InteractionEventCallback,
   ExtendedFeature,
 } from './use-map-interaction/index'
+export * from './use-login'
+export * from './use-login-redirect'
 export * from './use-map-interaction'
 export * from './use-dataviews-layers'
+export * from './use-local-storage'
 export * from './use-debounce'
 export * from './use-layer-composer'
 export * from './use-login'