+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @globalfishingwatchapp/fishing-map
 
+## 1.1.75
+
+### Patch Changes
+
+- Updated dependencies [08c49d086]
+  - @globalfishingwatch/layer-composer@10.1.1
+
+## 1.1.74
+
+### Patch Changes
+
+- Updated dependencies [ace3e03a1]
+  - @globalfishingwatch/react-hooks@15.3.0
+
 ## 1.1.73
 
 ### Patch Changes